refactor: Refactor UI components for improved readability and consistency
- Cleaned up code formatting in GraphDetailScreen, HistoryScreen, SettingsScreen, SyncSettingsScreen, WorkoutScreen, and SyncViewModel.
- Enhanced layout structure in HistoryScreen and SettingsScreen for better visual hierarchy.
- Updated text formatting in WorkoutScreen to display weight and reps in a more user-friendly manner.
- Removed unnecessary comments and whitespace for cleaner code.
- Added a new method in WorkoutViewModel to update sets, improving functionality.

+++ app/src/main/java/com/flandolf/workout/data/Entities.kt
@@ -36,7 +36,7 @@ data class ExerciseEntity(
     )], indices = [Index("exerciseId"), Index(value = ["firestoreId"], unique = true)]
 )
 data class SetEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val exerciseId: Long,
     val reps: Int,
     val weight: Float,
